refactor(build): extract Android app into :client:androidApp (AGP 9)
AGP 9 drops compatibility between `com.android.application` and the Kotlin
Multiplatform plugin (the KMP plugin warns that `com.android.application`
must move to a separate subproject). Split the Android application out of the
shared `:client:composeApp` KMP module:

- New `:client:androidApp` module applies `com.android.application` and holds
  the app entry points (MyApplication, MainActivity, AndroidApplicationComponent,
  the launcher manifest, res/, icons, signing config, applicationId/version). It
  depends on `:client:composeApp` for all shared code.
- `:client:composeApp` now applies `com.android.kotlin.multiplatform.library`
  instead of `com.android.application`; its Android target is configured via the
  new `androidLibrary { }` DSL (unique namespace `…chefmate.composeapp`,
  androidResources enabled, JVM 11). The `actual class BugsnagInitializer` stays
  here as part of the KMP library. Desktop (compose.desktop) and the iOS
  framework export are unchanged.
- `androidInstrumentedTest` source set renamed to the new plugin's
  `androidDeviceTest`.

Verified via Gradle: `:client:androidApp:assembleDebug` (APK builds, manifest
merges, Metro DI graph resolves across the new module boundary),
`:client:composeApp:compileKotlinJvm` (desktop), and the KMP/AGP-compat warning
is gone. NOT yet verified: on-device install and instrumented robot UI tests —
see the TODO in composeApp/build.gradle.kts; the commonTest -> androidDeviceTest
source-set routing still needs finalizing under the new plugin.
