+++ client/androidApp/src/main/AndroidManifest.xml
