+++ client/androidApp/src/main/res/mipmap-xxhdpi/ic_launcher_round.webp
