+++ client/androidApp/src/main/res/values/strings.xml
